silly test

+++ src/patternmodel.cpp
@@ -3558,7 +3558,7 @@ double GraphPatternModel::pmi(const EncAnyGram * key1, const EncAnyGram * key2) 
 double GraphPatternModel::npmi(const EncAnyGram * key1, const EncAnyGram * key2) {
     if ((rel_cooccurrences.count(key1)) && (rel_cooccurrences[key1].count(key2))) {
         const double jointcount = rel_cooccurrences[key1][key2];
-        return  log( (double) jointcount / (model->occurrencecount(key1) * model->occurrencecount(key2)) ) / -log((double)jointcount/(double)model->occurrences());    
+        return (double) log( (double) jointcount / (model->occurrencecount(key1) * model->occurrencecount(key2)) ) / (double) -log((double)jointcount/(double)model->occurrences());    
     } else {
         cerr << "Unable to compute NPMI, one of the operands does not exist" << endl;
         throw InternalError();
